+++ trinity/algorithm/advantage_fn/opmd_advantage.py
@@ -61,9 +61,11 @@ class OPMDAdvantageFn(AdvantageFn):
                     # TODO: consider id2baseline[idx] = id2score[idx] (so that this sample won't take effect?)
                 elif len(id2score[idx]) > 1:
                     if opmd_baseline == "mean":
-                        id2baseline[idx] = torch.mean(torch.tensor(id2score[idx]))
+                        id2baseline[idx] = torch.mean(
+                            torch.tensor(id2score[idx], dtype=torch.float32)
+                        )
                     elif opmd_baseline == "logavgexp":
-                        rewards_tensor = torch.tensor(id2score[idx])
+                        rewards_tensor = torch.tensor(id2score[idx], dtype=torch.float32)
                         # here we use the fact that logavgexp(x) = logsumexp(x) - log(len(x))
                         id2baseline[idx] = tau * (
                             torch.logsumexp(rewards_tensor / tau, dim=-1)
