+++ trinity/algorithm/advantage_fn/rloo_advantage.py
@@ -50,7 +50,7 @@ class RLOOAdvantageFn(AdvantageFn):
                 if len(id2score[idx]) == 1:
                     id2mean[idx] = torch.tensor(0.0)
                 elif len(id2score[idx]) > 1:
-                    id2mean[idx] = torch.mean(torch.tensor(id2score[idx]))
+                    id2mean[idx] = torch.mean(torch.tensor(id2score[idx], dtype=torch.float32))
                 else:
                     raise ValueError(f"no score in prompt index: {idx}")
             for i in range(bsz):
